requirements.txt
scipion-pyworkflow
scipion-em
scipion-chem
pandas
rdkit

